fix: show release button for all milestones
